ref: Rejigger config naming away from 'styles'
I'm giving the getOption part of the config module more responsibility
to generically translate whatever data is fed in. This allows the
configuration to share more code from Sentry in terms of what it does
with the data from the backend.

The main reason for this is so we don't have to duplicate things data
processing that the frontend might have been doing on charts into the
backend.

This also kills off (mostly) conflating 'style' with 'config'.