feat: fix memory leak when charts fail to render
Sometimes charts have invalid data passed to the chart options function. A 500 is expected, however each request is using more memory. Either from echarts.init being called and not disposed or it was erroring creating the echarts options which was after echarts.init, also causing dispose to not be called.