feat: upgrade echarts, canvas. Remove setCanvasCreator
I was just having some fun looking at the library alternatives, before we make a jump to another library. Try updating to the latest version of canvas and matching the echarts version we use in sentry.

We do not need to use setCanvasCreator since we are not adding any images into our graphs