ref: Use the VM to execute local config files
This makes config file loading consistent between using the http config
resolver and passing a local file path to load the config.

Consistency is important here for us since in Sentry we have
chartucterie acceptance tests, which are there to help us catch changes
made to Sentry's javascript app that may break chartucterie. However
they are run with chartucterie having loaded the config using a file
path and not HTTP, which differs from happens in production, where we
resolve the config via HTTP and execute it in a vm context.

By making these consistent we catch errors that happen loading the
config in the vm contextchartucterie VM.